==> src/00_setup.R <==

wd <- list()
wd$bin  <- "/Users/ccampbell/mammalToast/bin/"
wd$data <- "/Users/ccampbell/mammalToast/data/"
wd$figs <- "/Users/ccampbell/mammalToast/figs/"
wd$src  <- "/Users/ccampbell/mammalToast/src/"

require(plyr)
require(dplyr)
require(tidyr)
require(readxl)
require(purrr)
require(reshape2)
require(ggplot2)
